drug_screening_app/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pandas==2.2.3
numpy==1.26.4
rdkit-pypi==2022.9.5
scikit-learn==1.6.1
matplotlib==3.10.1
